Chirags update for journal and new admin menu

+++ admin/model/extension/extension.php
@@ -21,25 +21,9 @@ class ModelExtensionExtension extends Model {
 	}
 
 	public function uninstall($type, $code) {
+		if (strpos($code, 'journal2_') === 0) return;
+		
 		$this->db->query("DELETE FROM " . DB_PREFIX . "extension WHERE `type` = '" . $this->db->escape($type) . "' AND `code` = '" . $this->db->escape($code) . "'");
 		$this->db->query("DELETE FROM " . DB_PREFIX . "setting WHERE `code` = '" . $this->db->escape($code) . "'");
-	}
-	
-	public function createtable($query) { 
-		
-		$this->db->query($query);
-		
-	}
-	
-	public function Bypassusergroupcreate($name) { 
-		
-		$sql = "SELECT * FROM " . DB_PREFIX . "user_group where name = '" . $name . "' ";
-		$query = $this->db->query($sql);
-		$result = $query->rows;
-		if(empty($result))
-		{
-			$this->db->query("INSERT INTO " . DB_PREFIX . "user_group SET name = '" . $name . "', permission = ''");
-		}
-		
 	}
 }
